requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.2.0
django-cors-headers>=4.0.0
django-filter>=23.0
drf-yasg>=1.21.0
psycopg2-binary>=2.9.0
Pillow>=9.0.0
whitenoise>=6.0.0
django-extensions>=3.2.0
python-decouple>=3.6